+++ common/src/main/java/earth/terrarium/athena/api/client/utils/NullableEnumMap.java
@@ -153,8 +153,21 @@ public class NullableEnumMap<K extends Enum<K>, V> implements Map<K, V> {
 
         private final V value;
 
-        public MapEntry(V value) {
+        private MapEntry(V value) {
             this.value = value;
+        }
+
+        @Override
+        public boolean equals(Object obj) {
+            if (obj == this) return true;
+            if (obj == null || obj.getClass() != this.getClass()) return false;
+            //noinspection unchecked
+            return Objects.equals(this.value, ((MapEntry) obj).value);
+        }
+
+        @Override
+        public int hashCode() {
+            return Objects.hash(value);
         }
     }
 }